TST: Add shell script to run FileManager tests
