+++ .gitignore
@@ -51,6 +51,9 @@ tests/uploads/
 tests/drive/
 gramex/apps/filemanager/test/drive*
 gramex/apps/filemanager/test/tape.js
+gramex/apps/filemanager/test/nohup.out
+gramex/apps/filemanager/test/gramex-pid.txt
+
 
 # Don't commit ZIP files, except what's required for testing
 *.7z
