+++ .travis.yml
@@ -42,18 +42,19 @@ install:
   - pip install flake8 pep8-naming flake8-gramex flake8-blind-except flake8-print flake8-debugger bandit nose
   # Install node.js and packages using yarn (faster than npm)
   - npm install -g yarn
-  - yarn global add eclint eslint eslint-plugin-html htmllint-cli
+  - yarn global --ignore-engines add eclint eslint eslint-plugin-html htmllint-cli
   - yarn install
   # Set up variables
   - export BRANCH=$TRAVIS_BRANCH
 
 script:
   - eclint check '**/*.html' '**/*.js' '**/*.css' '**/*.yaml' '**/*.md'
-  - eslint gramex/apps
+  # - eslint gramex/apps
   - htmllint
   - flake8
   - bandit gramex --recursive --format csv || true
   - pip install -e .
   - pip install gramexenterprise
   - gramex license accept
-  - python setup.py nosetests
+  # - python setup.py nosetests
+  - cd gramex/apps/filemanager/ && sh tests/test.sh
