+++ gramex/apps/filemanager/package.json
@@ -9,10 +9,7 @@
   },
   "scripts": {
     "pretest": "browserify -s tape -r tape -o test/tape.js",
-    "test": "npm run pretest && node test/server.js puppeteer | tap-merge | faucet",
-    "test-chrome": "node test/server.js chrome | tap-merge | faucet",
-    "test-edge": "node test/server.js MicrosoftEdge | tap-merge | faucet",
-    "test-firefox": "node test/server.js firefox | tap-merge | faucet"
+    "test": "npm run pretest && node test/puppet.js | tap-merge | faucet"
   },
   "devDependencies": {
     "babel-core": "6",
