--- gramex/apps/filemanager/test/default-drive.json
@@ -1,1 +1,0 @@
-[{"id":1,"file":"imdb_ratings.csv","ext":".csv","path":"imdb-ratings.csv","size":791,"mime":"text\/csv","date":1589770947,"user_id":"jaidev.deshpande@gramener.com","user_role":null,"user_hd":"gramener.com","tag":""},{"id":2,"file":"actors.csv","ext":".csv","path":"actors.csv","size":443,"mime":"text\/csv","date":1589770947,"user_id":"jaidev.deshpande@gramener.com","user_role":null,"user_hd":"gramener.com","tag":""},{"id":3,"file":"data-list.json","ext":".json","path":"data-list.json","size":49,"mime":"application\/json","date":1589770974,"user_id":"jaidev.deshpande@gramener.com","user_role":null,"user_hd":"gramener.com","tag":""}]
