+++ gramex/apps/filemanager/test/test-filemanager.html
@@ -3,29 +3,29 @@
 
 <head>
   <title>filemanager tests</title>
-  <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
-  <link rel="stylesheet" href="../node_modules/font-awesome/css/font-awesome.min.css">
-  <script src="../node_modules/jquery/dist/jquery.min.js"></script>
-  <script src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="../node_modules/lodash/lodash.min.js"></script>
-  <script src="../node_modules/moment/min/moment.min.js"></script>
-  <script src="../node_modules/numeral/min/numeral.min.js"></script>
-  <script src="../node_modules/d3/build/d3.js"></script>
-  <script src="../node_modules/d3-scale-chromatic/dist/d3-scale-chromatic.js"></script>
-  <script src="../node_modules/g1/dist/g1.min.js"></script>
-  <script src="../node_modules/dropzone/dist/min/dropzone.min.js"></script>
-  <script src="../filemanager.js"></script>
-  <script src="tape.js"></script>
-  <script src="tape-stream.js"></script>
+  <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
+  <link rel="stylesheet" href="node_modules/font-awesome/css/font-awesome.min.css">
+  <script src="node_modules/jquery/dist/jquery.min.js"></script>
+  <script src="node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+  <script src="node_modules/lodash/lodash.min.js"></script>
+  <script src="node_modules/moment/min/moment.min.js"></script>
+  <script src="node_modules/numeral/min/numeral.min.js"></script>
+  <script src="node_modules/d3/build/d3.js"></script>
+  <script src="node_modules/d3-scale-chromatic/dist/d3-scale-chromatic.js"></script>
+  <script src="node_modules/g1/dist/g1.min.js"></script>
+  <script src="node_modules/dropzone/dist/min/dropzone.min.js"></script>
+  <script src="filemanager.js"></script>
+  <script src="test/tape.js"></script>
+  <script src="test/tape-stream.js"></script>
 </head>
 
 <body>
-  <div class="filemanager" data-src="/default"></div>
-  <div class="fm-opts" data-src="/default"></div>
-  <div class="fm-opts-2" data-src="/default"></div>
-  <div class="fm-opts-reserved" data-src="/default"></div>
-  <div class="edit-fname" data-src="/default"></div>
-  <div class="empty-drive" data-src="/drive1"></div>
+  <div class="filemanager" data-src="default"></div>
+  <div class="fm-opts" data-src="default"></div>
+  <div class="fm-opts-2" data-src="default"></div>
+  <div class="fm-opts-reserved" data-src="default"></div>
+  <div class="edit-fname" data-src="default"></div>
+  <div class="empty-drive" data-src="drive1"></div>
 </body>
 <script>
   tape('$().filemanager default', function(t) {
@@ -37,7 +37,7 @@
       t.deepEqual(column_names, _.map(e.options.columns, "name"))
 
       // check file links
-      let file_link = "/default?id=<%- row.id %>&_format=file&_download"
+      let file_link = "default?id=<%- row.id %>&_format=file&_download"
       t.equal(e.options.columns[0].link, file_link)
 
       // check delete button
